+++ src/main/java/kitchenpos/domain/order/Order.java
@@ -1,6 +1,5 @@
 package kitchenpos.domain.order;
 
-import kitchenpos.domain.OrderTable;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
