+++ src/main/java/kitchenpos/domain/order/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.tableGroup;
+package kitchenpos.domain.order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
