+++ src/test/java/kitchenpos/domain/order/OrderRepositoryTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.domain.order;
 
-import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.product.Product;
